projeto_programacao_web/requirements.txt
Flask==2.0.3
Jinja2==3.0.3
waitress==2.1.2
pyodbc==4.0.34
PyJWT==2.6.0
flask-login==0.6.2